refactor:order repository BigIntger->Long

+++ src/main/java/com/example/shoppingmallproject/order/repository/OrderRepository.java
@@ -4,8 +4,7 @@ import com.example.shoppingmallproject.order.entity.Orders;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Orders, BigInteger> {
+public interface OrderRepository extends JpaRepository<Orders, Long> {
 }
